Auth.sh

+++ auth.sh
@@ -1,21 +1,24 @@
 function auth {
 	APP_ID="12345"
 	URL="https://oauth.vk.com/authorize?client_id=$APP_ID&display=mobile&redirect_uri=https://oauth.vk.com/blank.html&scope=friends,photos,status,messages,wall,groups&response_type=token&v=$API_V"
-#TODO: check if w3m is avaible.
-	rm ./request.log
-	cp ~/.w3m/request.log ./request.log
-	w3m $URL -reqlog &
-	pid=$!
-	status=1
-	until [ $status -ne 1 ]; do
-		sleep .5
-		diff ~/.w3m/request.log ./request.log > diff.txt
-		grep "#access_token" ./diff.txt
-		status=$?
-		token=$(grep Location: ./diff.txt | grep "#access_token" | tail -1 | cut -c13-)
-        done
-	kill $pid
-	rm diff.txt
+	if [ -x "$(command -v w3m)" ]; then
+		rm ./request.log
+		cp ~/.w3m/request.log ./request.log
+		w3m $URL -reqlog &
+		pid=$!
+		status=1
+		until [ $status -ne 1 ]; do
+			sleep .5
+			diff ~/.w3m/request.log ./request.log > diff.txt
+			grep "#access_token" ./diff.txt
+			status=$?
+			token=$(grep Location: ./diff.txt | grep "#access_token" | tail -1 | cut -c13-)
+        	done
+		kill $pid
+		rm diff.txt
+	else
+		echo "w3m not found! Fallback mode is not implemented yet"
+	fi
 	#xdg-open $URL && echo "Copy and paste url here" && read token
 	expires_in=$(url.getvar "$token" expires_in)
 	expr $expires_in + $(date +%s)  > token.key
